- Restructured potentials.   - DynamicPotentialEvaluator is now implemented along the same lines as SphericalDynamicPotentialEvaluator   - Renamed the spherical potentials. They now have names like AngularKinetEnergyPotential and RadialKineticEnergyPotential. - Restructured pyste files   - in general, splitted existing pyste files into smaller units.   - separated aborber and potential pyste files   - separated example potentials from kinetic energy potentials. - Reduced the number of explicitly instantiated dimensions from 10 to 4. - Added more compilation options to allow compiling on AIX
git-svn-id: https://pyprop.googlecode.com/svn/trunk@46 354cffb3-6528-0410-8d42-45ce437ad3c5

+++ core/representation/lmrange.h
@@ -76,7 +76,7 @@ public:
 	//Gets the l-value from the linear index
 	static inline int GetL(int index)
 	{
-		return static_cast<int>(sqrt(index));		
+		return static_cast<int>(sqrt((double)index));		
 	}
 
 	//Gets the m-value from the linear index
